+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Features
+* cTrader | Add `requestedProtection` to orders [#103](https://github.com/Reiryoku-Technologies/Mida/pull/103)
+* Core | Add `requestedProtection` to `MidaOrder` for accessing an order's stop loss, take profit and trailing stop loss [#103](https://github.com/Reiryoku-Technologies/Mida/pull/103)
+
 2023.1.2 - 13-03-2023
 ===================
 ### Bug fixes